KS-2-4 the solution to lesson 2 task 4

+++ src/main/kotlin/Lesson2/lesson2-task4.kt
@@ -6,11 +6,11 @@ fun main () {
 
     var buff = 20
 
-    var buffCrystalOre = (crystalOre.toFloat() * buff.toFloat()) / 100
-    var buffIronOre = (ironOre.toFloat() * buff.toFloat()) / 100
+    var buffCrystalOre = (crystalOre * buff) / 100
+    var buffIronOre = (ironOre * buff) / 100
 
-    var finalCrystalOre = buffCrystalOre.toInt() + crystalOre
-    var finalIronOre = buffIronOre.toInt() + ironOre
+    var finalCrystalOre = buffCrystalOre + crystalOre
+    var finalIronOre = buffIronOre + ironOre
 
     println(finalCrystalOre)
     println(finalIronOre)
